html/semantics/sections/the-body-element/sections-body-events-content-manual.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Body Element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/single-page.html#the-body-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body onresize=myReaction.push("resize-content")
      onafterprint=myReaction.push("afterprint-content")
      onbeforeprint=myReaction.push("beforeprint-content")
      ononline=myReaction.push("online-content")
      onoffline=myReaction.push("offline-content")>

    <h1>Validation of body's content attributes requiring manual testing</h1>

    <p>Please do the following:</p>
    <ol>
        <li>resize this page</li>
        <li>print this page (to a file)</li>
        <li>choose to work offline</li>
        <li>come back online</li>
        <li>press the buttons immediately below this text</li>
    </ol>

    <form>
        <input type="button" id="btnResize" value="click me" />
        <input type="button" id="btnBefore" value="click me" />
        <input type="button" id="btnAfter" value="click me" />
        <input type="button" id="btnOnline" value="click me" />
        <input type="button" id="btnOffline" value="click me" />
    </form>

    <div id="log"></div>

    <script>
        "use strict";

        var myReaction = [], testResize = {}, testBeforePrint = {}, testAfterPrint = {}, testOnline = {}, testOffline = {};

        // need to be able to wait for user to push button
        setup({ explicit_timeout: true });

        // register async tests
        testResize = async_test("Check body's onresize content attribute"); 
        testBeforePrint = async_test("Check body's onbeforeprint content attribute");
        testAfterPrint = async_test("Check body's onafterprint content attribute");
        testOnline = async_test("Check body's ononline content attribute");
        testOffline = async_test("Check body's onoffline content attribute");

        // run async tests after buttons are clicked
        document.getElementById("btnResize").onclick = testResize.step_func(function (event) {     	
                assert_in_array("resize-content", myReaction, "onresize content attribute failed: ");
                testResize.done();
        });
        document.getElementById("btnBefore").onclick = testBeforePrint.step_func(function (event) {     	
                assert_in_array("beforeprint-content", myReaction, "onbeforeprint content attribute failed: ");
                testBeforePrint.done();
        });
        document.getElementById("btnAfter").onclick = testAfterPrint.step_func(function (event) {      	
                assert_in_array("afterprint-content", myReaction, "onafterprint content attribute failed: ");
                testAfterPrint.done();
    	});
        document.getElementById("btnOnline").onclick = testOnline.step_func(function (event) {     	
                assert_in_array("online-content", myReaction, "ononline content attribute failed: ");
                testOnline.done();
        });
        document.getElementById("btnOffline").onclick = testOffline.step_func(function (event) {      	
                assert_in_array("offline-content", myReaction, "onoffline content attribute failed: ");
                testOffline.done();
        })

    </script>

</body>
